IntelliTour_Conversational_Tourism_Chatbot/requirements.txt
#In order to ensure smooth running of the project, install the following libraries
flask
python-dotenv
openai
aiohttp
requests
amadeus